Some more updates
In the future, an update I am thinking of is that perhaps I can make a virtual local environment where youtube-dl can be installed in its most recent version and use that space to download songs/videos. So everything can be sort of isolated in one part for security reasons as well. Just a fun thought.